1. Moved KickUser code from Atomist into src/gluon/util/team/Teams.ts 2. Added try catch around removeMemberFromChannel so that the event handle continues running and removes user from projects etc 3. Updated error handling on src/gluon/services/team/RemoveMemberFromTeamService.ts removeUserFromSlackChannel
